поправил БД + сервисы

+++ src/main/java/com/example/WalletProject/repositories/ClientRepository.java
@@ -4,8 +4,7 @@ import com.example.WalletProject.models.Client;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigInteger;
 //в параметрах репозитория указаны: тип того, что будем искать (Account к примеру), тип, по которому будем искать (к примеру Account Id/AuthId)
 @Repository
-public interface ClientRepository extends JpaRepository<Client, BigInteger> {
+public interface ClientRepository extends JpaRepository<Client, Long> {
 }
